+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add app::set_raw_callback() which avoids boxing callbacks.
 - Update FLTK to commit cd26829 which adds MacOS 11.0 support.
 - Add missing docs for callback arguments.
+- Add TextBuffer::save_file().
 
 ## [0.8.0] - 2020-08-02
 ### Changes